\name{LCstrategy-internal}
\alias{LCstrategy-internal}
\alias{plot.LCcluster}
\alias{plot.confirm}
\alias{plot.KSperm}
\alias{print.LCcluster}
\alias{print.ivadj}
\alias{print.lrcagg}
\alias{print.ltdagg}
\alias{print.confirm}
\alias{print.KSperm}
\title{Internal LocalControlStrategy functions.}
\description{Internal functions for the LocalControlStrategy-package.  Plot and print
 functions require the output object from a LocalControlStrategy function as their
 first argument.}
\author{Bob Obenchain <wizbob@att.net>}
\keyword{methods}
\keyword{internal}
